start js client
