+++ root/side_modules/number.py
@@ -47,3 +47,5 @@ if __name__ == '__main__':
     #     [14706,17085,24848]
     #     ]
     # }
+    #
+    # {"B": [[31852, 16866, 13534]]}
