+++ root/static/style.css
@@ -1,0 +1,3 @@
+.selector-for-some-widget {
+    box-sizing: content-box;
+  }
