+++ root/static/websockets_index.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <title>Connect Four</title>
+  </head>
+  <body>
+    
+    <div class="main">
+      <h1>WebSockets Page</h1>
+    </div>
+
+    <script src="websockets_main.js" type="module"></script>
+  
+  </body>
+</html>
